Moved index page to frontpage.
